toolkit/content/tests/widgets/test_videocontrols_audio.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Video controls with Audio file test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>

<div id="content">
  <video id="video" controls preload="metadata"></video>
</div>

<pre id="test">
<script class="testbody" type="application/javascript">

  const video = document.getElementById("video");

  function findElementByAttribute(element, aName, aValue) {
    if (!("getAttribute" in element)) {
      return false;
    }
    if (element.getAttribute(aName) === aValue) {
      return element;
    }
    let children =
      InspectorUtils.getChildrenForNode(element, true);
    for (let child of children) {
      var result = findElementByAttribute(child, aName, aValue);
      if (result) {
        return result;
      }
    }
    return false;
  }

  function loadedmetadata(event) {
    SimpleTest.requestCompleteLog();
    SimpleTest.executeSoon(async function test_fullscreen_unavailable() {
      const { x, y } = video.getBoundingClientRect();
      const controlBar = findElementByAttribute(video, "class", "controlBar");
      SimpleTest.requestFlakyTimeout("Waiting to ensure that fullscreen event does not fire");
      const endedPromise = new Promise(resolve => {
        video.addEventListener("ended", () => {
          info('Video ended event fired!');
          resolve();
        }, { once: true });
        setTimeout( () => {
          info('Video ran out of time before ended event fired!');
          resolve();
        }, video.duration * 1000);
      });
      let noFullscreenEvent = true;
      is(controlBar.getAttribute("fullscreen-unavailable"), "true", "Fullscreen button is hidden");
      document.addEventListener("mozfullscreenchange", () => {
        noFullscreenEvent = false;
      }, { once: true });
      info("Simulate double click on media player.");
      synthesizeMouse(video, x, y, { clickCount: 2 });
      info("Waiting for video to end...");
      await endedPromise;
      // By this point, if the double click was going to trigger fullscreen then
      // it should have happened by now.
      ok(
        noFullscreenEvent,
        "Double clicking should not trigger fullscreen event"
      );
      SimpleTest.finish();
    });
  }

  SpecialPowers.pushPrefEnv({"set": [["media.cache_size", 40000]]}, startTest);
  function startTest() {
    // Kick off test once audio has loaded.
    video.addEventListener("loadedmetadata", loadedmetadata, { once: true });
    video.src = "audio.ogg";
  }

  SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
